# src/common/client/openapi.json
{
  "openapi": "3.0.2",
  "info": { "title": "Druffle Development API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/admins/{admin_id}/info": {
      "get": {
        "tags": ["Admins: Info"],
        "summary": "Get Info",
        "description": "Get admin info by id. Applicable only for Admin and Druffler user only",
        "operationId": "get_info_api_v1_admins__admin_id__info_get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Admin Id", "type": "string" },
            "name": "admin_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AdminInfoResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "put": {
        "tags": ["Admins: Info"],
        "summary": "Update Admin Info By Admin Id",
        "description": "Update admin info by admin id. Applicable only for Admin and Druffler user only",
        "operationId": "update_admin_info_by_admin_id_api_v1_admins__admin_id__info_put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Admin Id", "type": "string" },
            "name": "admin_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminInfoUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AdminInfoResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
...
    "/api/v1/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get All Accounts",
        "description": "Get all accounts. Applicable only for Druffler user only",
        "operationId": "get_all_accounts_api_v1_accounts_get",
        "parameters": [
          {
            "required": false,
            "schema": { "title": "Skip", "type": "integer", "default": 0 },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": { "title": "Limit", "type": "integer", "default": 50 },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountsListResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "Create Account",
        "description": "Create new account. Applicable only for Druffler user only",
        "operationId": "create_account_api_v1_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountCreateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AccountResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
    "/api/v1/accounts/{account_id}": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get Accounts By Id",
        "description": "Update account info. Applicable only for Administrative user only",
        "operationId": "get_accounts_by_id_api_v1_accounts__account_id__get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Account Id", "type": "string" },
            "name": "account_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AccountResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "put": {
        "tags": ["Accounts"],
        "summary": "Update Accounts",
        "description": "Update account info. Applicable only for Administrative user only",
        "operationId": "update_accounts_api_v1_accounts__account_id__put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Account Id", "type": "string" },
            "name": "account_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AccountResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "delete": {
        "tags": ["Accounts"],
        "summary": "Delete Account By Id",
        "description": "Delete account by id. Applicable for Druffler accounts",
        "operationId": "delete_account_by_id_api_v1_accounts__account_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Account Id", "type": "string" },
            "name": "account_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StatusResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
    "/api/v1/accounts/{account_id}/admins": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get All Admins By Account Id",
        "description": "Get all admins. Applicable only for Druffler user only",
        "operationId": "get_all_admins_by_account_id_api_v1_accounts__account_id__admins_get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Account Id", "type": "string" },
            "name": "account_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": { "title": "Skip", "type": "integer", "default": 0 },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": { "title": "Limit", "type": "integer", "default": 50 },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AdminListResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "Create Admin",
        "description": "Create new administrator. Applicable only for Druffler user",
        "operationId": "create_admin_api_v1_accounts__account_id__admins_post",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Account Id", "type": "string" },
            "name": "account_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AdminUpdateRequest" } }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AdminFullResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
...
    "/api/v1/hotels/{hotel_id}/accommodations": {
      "get": {
        "tags": ["Hotel: Accommodations"],
        "summary": "Get Accommodations By Hotel Id",
        "description": "Get accommodations by hotel id. Applicable for Administrative accounts",
        "operationId": "get_accommodations_by_hotel_id_api_v1_hotels__hotel_id__accommodations_get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelAccommodationsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "put": {
        "tags": ["Hotel: Accommodations"],
        "summary": "Update Accommodations By Hotel Id",
        "description": "Get accommodations by hotel id. Applicable for Administrative accounts",
        "operationId": "update_accommodations_by_hotel_id_api_v1_hotels__hotel_id__accommodations_put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelAccommodationsUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelAccommodationsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      },
      "post": {
        "tags": ["Hotel: Accommodations"],
        "summary": "Create Accommodations",
        "description": "Create new hotel accommodations. Applicable for Administrative account",
        "operationId": "create_accommodations_api_v1_hotels__hotel_id__accommodations_post",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Hotel Id", "type": "string" },
            "name": "hotel_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelAccommodationsUpdateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotelAccommodationsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "JWTokenAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreateRequest": {
        "title": "AccountCreateRequest",
        "required": ["contact_name", "contact_phone"],
        "type": "object",
        "properties": {
          "name": { "title": "Name", "type": "string" },
          "description": { "title": "Description", "type": "string" },
          "contact_name": { "title": "Contact Name", "type": "string" },
          "contact_phone": { "title": "Contact Phone", "type": "string" },
          "is_active": { "title": "Is Active", "type": "boolean", "default": false },
          "agreement": { "title": "Agreement", "type": "string" }
        }
      },
      "AccountResponse": {
        "title": "AccountResponse",
        "required": ["contact_name", "contact_phone", "id"],
        "type": "object",
        "properties": {
          "is_deleted": { "title": "Is Deleted", "type": "boolean", "default": false },
          "name": { "title": "Name", "type": "string" },
          "description": { "title": "Description", "type": "string" },
          "contact_name": { "title": "Contact Name", "type": "string" },
          "contact_phone": { "title": "Contact Phone", "type": "string" },
          "is_active": { "title": "Is Active", "type": "boolean", "default": false },
          "agreement": { "title": "Agreement", "type": "string" },
          "id": { "title": "Id", "type": "string", "format": "uuid4" }
        }
      },
      "AccountUpdateRequest": {
        "title": "AccountUpdateRequest",
        "type": "object",
        "properties": {
          "name": { "title": "Name", "type": "string" },
          "description": { "title": "Description", "type": "string" },
          "contact_name": { "title": "Contact Name", "type": "string" },
          "contact_phone": { "title": "Contact Phone", "type": "string" },
          "is_active": { "title": "Is Active", "type": "boolean", "default": false }
        }
      },
      "AccountsListResponse": {
        "title": "AccountsListResponse",
        "required": ["data", "count"],
        "type": "object",
        "properties": {
          "data": {
            "title": "Data",
            "type": "array",
            "items": { "$ref": "#/components/schemas/AccountResponse" }
          },
          "count": { "title": "Count", "type": "integer" }
        }
      },
...
      "AdminFullResponse": {
        "title": "AdminFullResponse",
        "required": [
          "id",
          "username",
          "admin_type",
          "is_active",
          "is_deleted",
          "created_at",
          "updated_at",
          "info"
        ],
        "type": "object",
        "properties": {
          "role": { "title": "Role", "type": "string" },
          "id": { "title": "Id", "type": "string", "format": "uuid4" },
          "username": { "title": "Username", "type": "string", "format": "email" },
          "admin_type": { "$ref": "#/components/schemas/AdminTypes" },
          "is_active": { "title": "Is Active", "type": "boolean" },
          "is_deleted": { "title": "Is Deleted", "type": "boolean" },
          "created_at": { "title": "Created At", "type": "string", "format": "date-time" },
          "updated_at": { "title": "Updated At", "type": "string", "format": "date-time" },
          "info": { "$ref": "#/components/schemas/AdminInfoResponse" }
        }
      },
      "AdminInfoResponse": {
        "title": "AdminInfoResponse",
        "required": ["phone"],
        "type": "object",
        "properties": {
          "first_name": { "title": "First Name", "type": "string" },
          "last_name": { "title": "Last Name", "type": "string" },
          "email": { "title": "Email", "type": "string" },
          "phone": { "title": "Phone", "type": "string" },
          "position": { "title": "Position", "type": "string" },
          "description": { "title": "Description", "type": "string" },
          "birthday": { "title": "Birthday", "type": "string", "format": "date-time" }
        }
      },
      "AdminInfoUpdateRequest": {
        "title": "AdminInfoUpdateRequest",
        "required": ["phone"],
        "type": "object",
        "properties": {
          "first_name": { "title": "First Name", "type": "string" },
          "last_name": { "title": "Last Name", "type": "string" },
          "email": { "title": "Email", "type": "string" },
          "phone": { "title": "Phone", "type": "string" },
          "position": { "title": "Position", "type": "string" },
          "description": { "title": "Description", "type": "string" },
          "birthday": { "title": "Birthday", "type": "string", "format": "date-time" }
        }
      },
      "AdminListResponse": {
        "title": "AdminListResponse",
        "required": ["data", "count"],
        "type": "object",
        "properties": {
          "data": {
            "title": "Data",
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdminResponse" }
          },
          "count": { "title": "Count", "type": "integer" }
        }
      },
      "AdminResponse": {
        "title": "AdminResponse",
        "required": [
          "id",
          "username",
          "admin_type",
          "is_active",
          "is_deleted",
          "created_at",
          "updated_at"
        ],
        "type": "object",
        "properties": {
          "role": { "title": "Role", "type": "string" },
          "id": { "title": "Id", "type": "string", "format": "uuid4" },
          "username": { "title": "Username", "type": "string", "format": "email" },
          "admin_type": { "$ref": "#/components/schemas/AdminTypes" },
          "is_active": { "title": "Is Active", "type": "boolean" },
          "is_deleted": { "title": "Is Deleted", "type": "boolean" },
          "created_at": { "title": "Created At", "type": "string", "format": "date-time" },
          "updated_at": { "title": "Updated At", "type": "string", "format": "date-time" }
        }
      },
      "AdminTypes": {
        "title": "AdminTypes",
        "enum": ["Administrator", "Employee"],
        "description": "An enumeration."
      },
      "AdminUpdateRequest": {
        "title": "AdminUpdateRequest",
        "required": ["admin_type", "username", "info"],
        "type": "object",
        "properties": {
          "admin_type": { "$ref": "#/components/schemas/AdminTypes" },
          "username": { "title": "Username", "type": "string", "format": "email" },
          "info": { "$ref": "#/components/schemas/AdminInfoUpdateRequest" },
          "hotel_id": { "title": "Hotel Id", "type": "string", "format": "uuid4" }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
...
      "HotelAccommodationsResponse": {
        "title": "HotelAccommodationsResponse",
        "required": [
          "check_in",
          "check_out",
          "children_allowed",
          "home_pet_allowed",
          "hotel_id",
          "id"
        ],
        "type": "object",
        "properties": {
          "check_in": { "title": "Check In", "type": "string", "format": "time" },
          "check_out": { "title": "Check Out", "type": "string", "format": "time" },
          "breakfast_start": { "title": "Breakfast Start", "type": "string", "format": "time" },
          "breakfast_end": { "title": "Breakfast End", "type": "string", "format": "time" },
          "weekend_breakfast_start": {
            "title": "Weekend Breakfast Start",
            "type": "string",
            "format": "time"
          },
          "weekend_breakfast_end": {
            "title": "Weekend Breakfast End",
            "type": "string",
            "format": "time"
          },
          "children_allowed": { "title": "Children Allowed", "type": "boolean" },
          "children_age_range": { "title": "Children Age Range", "type": "string" },
          "children_cost": { "title": "Children Cost", "type": "string" },
          "home_pet_allowed": { "title": "Home Pet Allowed", "type": "boolean" },
          "pet_weight": { "title": "Pet Weight", "type": "string" },
          "pet_deposit": { "title": "Pet Deposit", "type": "string" },
          "hotel_id": { "title": "Hotel Id", "type": "string", "format": "uuid4" },
          "id": { "title": "Id", "type": "string", "format": "uuid4" }
        }
      },
      "HotelAccommodationsUpdateRequest": {
        "title": "HotelAccommodationsUpdateRequest",
        "required": ["check_in", "check_out", "children_allowed", "home_pet_allowed"],
        "type": "object",
        "properties": {
          "check_in": { "title": "Check In", "type": "string", "format": "time" },
          "check_out": { "title": "Check Out", "type": "string", "format": "time" },
          "breakfast_start": { "title": "Breakfast Start", "type": "string", "format": "time" },
          "breakfast_end": { "title": "Breakfast End", "type": "string", "format": "time" },
          "weekend_breakfast_start": {
            "title": "Weekend Breakfast Start",
            "type": "string",
            "format": "time"
          },
          "weekend_breakfast_end": {
            "title": "Weekend Breakfast End",
            "type": "string",
            "format": "time"
          },
          "children_allowed": { "title": "Children Allowed", "type": "boolean" },
          "children_age_range": { "title": "Children Age Range", "type": "string" },
          "children_cost": { "title": "Children Cost", "type": "string" },
          "home_pet_allowed": { "title": "Home Pet Allowed", "type": "boolean" },
          "pet_weight": { "title": "Pet Weight", "type": "string" },
          "pet_deposit": { "title": "Pet Deposit", "type": "string" }
        }
      },
...
      "StatusResponse": {
        "title": "StatusResponse",
        "required": ["status"],
        "type": "object",
        "properties": {
          "status": { "title": "Status", "enum": ["ok", "fail"], "type": "string" },
          "details": { "title": "Details", "type": "string" }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "JWTokenAuth": { "type": "apiKey", "in": "header", "name": "Authorization" }
    }
  }
}